requirements.txt
# 核心依赖
PyQt5>=5.15.0
ffmpeg-python>=0.2.0
openai-whisper>=20230308
pysrt>=1.1.2
requests>=2.28.1
numpy>=1.23.5
pydub>=0.25.1
librosa>=0.10.0
psutil>=5.9.0

# 可选依赖
torch>=2.0.0  # 如需GPU加速，建议安装CUDA兼容版本
transformers>=4.28.0  # 可选，用于替代翻译API

# Windows特定依赖
pywin32>=306; platform_system == "Windows"

# macOS特定依赖
pyobjc>=9.0; platform_system == "Darwin"

# 开发依赖
pytest>=7.3.1
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
